refactor: add layout for pages
