+++ src/components/Layout.tsx
@@ -1,0 +1,12 @@
+import Navigation from "./Navigation";
+
+const Layout = ({ children }: { children: React.ReactNode }) => {
+  return (
+    <>
+      <Navigation />
+      {children}
+    </>
+  );
+};
+
+export default Layout;
